+++ app/src/main/java/com/example/frume/data/Review.kt
@@ -1,5 +1,7 @@
 package com.example.frume.data
 
+import android.net.Uri
+
 data class Review(
-    val reviewImage: String?
+    val reviewImage: Uri?
 )
