+++ app/src/main/java/com/example/frume/fragment/user_fragment/user_cart/UserCartFragmentSubscribe.kt
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [UserCartFragment2.newInstance] factory method to
+ * Use the [UserCartFragmentSubscribe.newInstance] factory method to
  * create an instance of this fragment.
  */
-class UserCartFragment2 : Fragment() {
+class UserCartFragmentSubscribe : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class UserCartFragment2 : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_user_cart2, container, false)
+        return inflater.inflate(R.layout.fragment_user_cart_subscribe, container, false)
     }
 
     companion object {
@@ -50,7 +50,7 @@ class UserCartFragment2 : Fragment() {
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            UserCartFragment2().apply {
+            UserCartFragmentSubscribe().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
