+++ app/src/main/res/anim/slide_out_right.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+
+	<translate android:fromXDelta="0%" android:toXDelta="100%"
+		android:fromYDelta="0%" android:toYDelta="0%"
+		android:duration="500"/>
+</set>
