+++ app/src/main/res/drawable/background_remove.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+	android:shape="rectangle">
+	<solid android:color="#80D9D9D9" />
+	<corners android:radius="12dp" />
+</shape>
