--- app/src/main/res/layout/fragment_user_cart1.xml
@@ -1,167 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-
-	xmlns:app="http://schemas.android.com/apk/res-auto"
-	xmlns:tools="http://schemas.android.com/tools">
-
-	<androidx.core.widget.NestedScrollView
-		android:layout_width="match_parent"
-		android:layout_height="match_parent"
-		android:fillViewport="true">
-
-		<androidx.constraintlayout.widget.ConstraintLayout
-			android:layout_width="match_parent"
-			android:layout_height="match_parent">
-
-			<TextView
-				android:id="@+id/textViewUserCartUserName"
-				style="@style/AppText3.S14"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:layout_margin="20dp"
-				app:layout_constraintStart_toStartOf="parent"
-				app:layout_constraintTop_toTopOf="parent" />
-
-			<TextView
-				android:id="@+id/textViewUserCartUserAddress"
-				style="@style/AppText3.S14"
-				android:layout_width="0dp"
-				android:layout_height="wrap_content"
-				android:layout_marginTop="10dp"
-				android:layout_marginEnd="20dp"
-				android:maxLines="2"
-				app:drawableStartCompat="@drawable/ic_location_on_24px"
-				app:layout_constraintEnd_toEndOf="parent"
-				app:layout_constraintStart_toStartOf="@id/textViewUserCartUserName"
-				app:layout_constraintTop_toBottomOf="@id/textViewUserCartUserName" />
-
-			<TextView
-				android:id="@+id/textViewUserCartUserPhoneNumber"
-				style="@style/AppText3.S14"
-				android:layout_width="0dp"
-				android:layout_height="wrap_content"
-				android:layout_marginTop="10dp"
-				app:drawableStartCompat="@drawable/ic_call_24px"
-				app:layout_constraintStart_toStartOf="@id/textViewUserCartUserAddress"
-				app:layout_constraintTop_toBottomOf="@id/textViewUserCartUserAddress" />
-
-			<androidx.appcompat.widget.AppCompatButton
-				android:id="@+id/buttonUserCartDialogModifyAddress"
-				style="@style/AppText3.S14"
-				android:layout_width="0dp"
-				android:layout_height="wrap_content"
-				android:layout_marginHorizontal="20dp"
-				android:layout_marginTop="20dp"
-				android:background="@drawable/background_input_count"
-				android:text="배송지 변경"
-				app:layout_constraintEnd_toEndOf="parent"
-				app:layout_constraintStart_toStartOf="parent"
-				app:layout_constraintTop_toBottomOf="@id/textViewUserCartUserPhoneNumber" />
-			<!--전체구매 체크박스 -->
-			<com.google.android.material.checkbox.MaterialCheckBox
-				android:id="@+id/checkboxUserCartSelectAll"
-				style="@style/AppText3.S14"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:layout_margin="15sp"
-				android:text="전체 선택"
-				app:layout_constraintStart_toStartOf="parent"
-				app:layout_constraintTop_toBottomOf="@id/buttonUserCartDialogModifyAddress" />
-
-			<TextView
-				android:id="@+id/textViewButtonUserCartDelete"
-				style="@style/AppText3.S14"
-				android:layout_width="0dp"
-				android:layout_height="wrap_content"
-				android:layout_marginEnd="20dp"
-				android:text="선택삭제"
-				app:layout_constraintBottom_toBottomOf="@id/checkboxUserCartSelectAll"
-				app:layout_constraintEnd_toEndOf="parent"
-				app:layout_constraintTop_toBottomOf="@id/buttonUserCartDialogModifyAddress"
-				app:layout_constraintTop_toTopOf="@id/checkboxUserCartSelectAll" />
-
-			<androidx.recyclerview.widget.RecyclerView
-				android:id="@+id/recycler_view_user_cart1"
-				android:layout_width="0dp"
-				android:layout_height="match_parent"
-				android:layout_margin="10dp"
-				android:nestedScrollingEnabled="false"
-				app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-				app:layout_constraintEnd_toEndOf="parent"
-				app:layout_constraintStart_toStartOf="parent"
-				app:layout_constraintTop_toBottomOf="@id/checkboxUserCartSelectAll"
-				tools:listitem="@layout/item_usercart_list" />
-
-			<TextView
-				android:id="@+id/textViewBottomSheetShowCartOptionDateLabel"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:layout_marginTop="30dp"
-				android:fontFamily="@font/nanumsquareneo_5"
-				android:text="배송 예정일 선택"
-				app:layout_constraintStart_toStartOf="@id/textViewUserCartUserName"
-				app:layout_constraintTop_toBottomOf="@+id/recycler_view_user_cart1" />
-
-			<View
-				android:id="@+id/viewBottomSheetShowCartOptionDate"
-				android:layout_width="0dp"
-				android:layout_height="60dp"
-				android:layout_marginHorizontal="20dp"
-				android:layout_marginTop="20dp"
-				android:background="@drawable/background_input_count"
-				app:layout_constraintEnd_toEndOf="parent"
-				app:layout_constraintStart_toStartOf="parent"
-				app:layout_constraintTop_toBottomOf="@id/textViewBottomSheetShowCartOptionDateLabel" />
-
-			<TextView
-				android:id="@+id/textViewBottomSheetShowCartOptionDate"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:fontFamily="@font/nanumsquareneo_3"
-				android:text="2025년 01월 14일 (화)"
-				android:textSize="18dp"
-				app:layout_constraintBottom_toBottomOf="@id/viewBottomSheetShowCartOptionDate"
-				app:layout_constraintEnd_toEndOf="@id/viewBottomSheetShowCartOptionDate"
-				app:layout_constraintStart_toStartOf="@id/viewBottomSheetShowCartOptionDate"
-				app:layout_constraintTop_toTopOf="@id/viewBottomSheetShowCartOptionDate" />
-
-			<TextView
-				android:id="@+id/textViewUserCartDialogPriceLabel"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:layout_marginLeft="20dp"
-				android:layout_marginTop="60dp"
-				android:fontFamily="@font/nanumsquareneo_3"
-				android:text="총 결제 예상 금액"
-				android:textSize="20dp"
-				app:layout_constraintStart_toStartOf="@id/textViewBottomSheetShowCartOptionDateLabel"
-				app:layout_constraintTop_toBottomOf="@id/viewBottomSheetShowCartOptionDate" />
-
-			<TextView
-				android:id="@+id/textViewUserCartDialogPrice"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:layout_marginTop="60dp"
-				android:layout_marginEnd="40dp"
-				android:fontFamily="@font/nanumsquareneo_3"
-				android:text="15,000원"
-				android:textSize="20dp"
-				app:layout_constraintEnd_toEndOf="parent"
-				app:layout_constraintTop_toBottomOf="@id/viewBottomSheetShowCartOptionDate" />
-
-			<androidx.appcompat.widget.AppCompatButton
-				android:id="@+id/buttonUserCartOrder"
-				style="@style/CustomButtonStyle"
-				android:layout_width="0dp"
-				android:layout_height="wrap_content"
-				android:layout_marginHorizontal="20dp"
-				android:layout_marginTop="50dp"
-				android:layout_marginBottom="40dp"
-				android:text="주문하기"
-				app:layout_constraintBottom_toBottomOf="parent"
-				app:layout_constraintEnd_toEndOf="parent"
-				app:layout_constraintStart_toStartOf="parent"
-				app:layout_constraintTop_toBottomOf="@id/textViewUserCartDialogPriceLabel" />
-		</androidx.constraintlayout.widget.ConstraintLayout>
-	</androidx.core.widget.NestedScrollView>
-</layout>
