+++ app/src/main/res/layout/fragment_user_cart_subscribe.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.user_fragment.user_cart.UserCartFragment2">
+    tools:context=".fragment.user_fragment.user_cart.UserCartFragmentSubscribe">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
