+++ app/src/main/res/layout/fragment_user_order_history.xml
@@ -21,13 +21,12 @@
 
 		<TextView
 			android:id="@+id/textViewUserOrderHistoryOrderFilterText"
+			style="@style/AppText5.S20"
 			android:layout_width="wrap_content"
 			android:layout_height="wrap_content"
 			android:layout_marginStart="40dp"
 			android:layout_marginTop="45dp"
-			android:fontFamily="@font/nanumsquareneo_5"
 			android:text="주문 상태"
-			android:textSize="20dp"
 			app:layout_constraintEnd_toStartOf="@id/textInputLayoutUserOrderHistoryOrderStateFilter"
 			app:layout_constraintStart_toStartOf="parent"
 			app:layout_constraintTop_toBottomOf="@id/toolbarUserOrderHistory" />
@@ -37,7 +36,6 @@
 			android:layout_width="wrap_content"
 			android:layout_height="wrap_content"
 			android:entries="@array/item_user_order_state_order_state"
-			android:gravity="center"
 			app:endIconMode="dropdown_menu"
 			app:layout_constraintBottom_toBottomOf="@id/textViewUserOrderHistoryOrderFilterText"
 			app:layout_constraintEnd_toEndOf="parent"
@@ -56,12 +54,11 @@
 
 		<TextView
 			android:id="@+id/textViewUserOrderHistorySearchPeriod"
+			style="@style/AppText5.S20"
 			android:layout_width="wrap_content"
 			android:layout_height="wrap_content"
 			android:layout_marginTop="45dp"
-			android:fontFamily="@font/nanumsquareneo_5"
 			android:text="조회 기간"
-			android:textSize="20dp"
 			app:layout_constraintStart_toStartOf="@id/textViewUserOrderHistoryOrderFilterText"
 			app:layout_constraintTop_toBottomOf="@id/textViewUserOrderHistoryOrderFilterText" />
 
@@ -82,8 +79,7 @@
 				android:entries="@array/item_user_order_state_order_state"
 				android:gravity="center"
 				android:hint="조회 기간"
-				android:inputType="none"
-				/>
+				android:inputType="none" />
 		</com.google.android.material.textfield.TextInputLayout>
 
 		<com.google.android.material.divider.MaterialDivider
