+++ app/src/main/res/layout/item_add_image.xml
@@ -38,6 +38,7 @@
 				android:layout_width="wrap_content"
 				android:layout_height="wrap_content"
 				android:layout_gravity="end|top"
+				android:background="@drawable/background_remove"
 				android:layout_margin="5dp"
 				android:src="@drawable/ic_cancel" />
 		</com.google.android.material.card.MaterialCardView>
