+++ app/src/main/res/values/colors.xml
@@ -1,22 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="transparent">#00000000</color>
-    <color name="black">#FF000000</color>
-    <color name="white">#FFFFFFFF</color>
+	<color name="transparent">#00000000</color>
+	<color name="black">#FF000000</color>
+	<color name="white">#FFFFFFFF</color>
 
-    <color name="green10">#E7FFEB</color>
-    <color name="green50">#75E7A4</color>
-    <color name="green100">#03CF5D</color>
+	<color name="green10">#E7FFEB</color>
+	<color name="green50">#75E7A4</color>
+	<color name="green100">#03CF5D</color>
 
-    <color name="gray5">#f2f2f2</color>
-    <color name="gray8">#f0f0f0</color>
-    <color name="gray10">#d9d9d9</color>
-    <color name="gray50">#b9b9b9</color>
-    <color name="gray100">#999999</color>
-
-
-    <color name="yellow100">#FFD54F</color>
+	<color name="gray5">#f2f2f2</color>
+	<color name="gray8">#f0f0f0</color>
+	<color name="gray10">#d9d9d9</color>
+	<color name="gray15">#d3d3d3</color>
+	<color name="gray50">#b9b9b9</color>
+	<color name="gray100">#999999</color>
 
 
-    <color name="red50">#FA7070</color>
+	<color name="yellow100">#FFD54F</color>
+
+
+	<color name="red50">#FA7070</color>
 </resources>
