+++ app/src/main/res/values/styles.xml
@@ -51,6 +51,10 @@
 		<item name="android:textColor">@color/black</item>
 	</style>
 
+	<style name="AppText3.S12">
+		<item name="android:textSize">12sp</item>
+	</style>
+
 	<style name="AppText3.S14">
 		<item name="android:textSize">14sp</item>
 	</style>
@@ -67,6 +71,10 @@
 	<style name="AppText4">
 		<item name="fontFamily">@font/nanumsquareneo_4</item>
 		<item name="android:textColor">@color/black</item>
+	</style>
+
+	<style name="AppText4.S12">
+		<item name="android:textSize">12sp</item>
 	</style>
 
 	<style name="AppText4.S14">
